+++ tests/switchCase.c
@@ -1,6 +1,6 @@
 int main(){
     int a = 10, b = 343;
-    float f = 10.43783,  g = 23.23;
+    double f = 10.43783,  g = 23.23;
     printf("f = %lf\n", f);
     switch(a){
         case 10:{
